refactor: rename credit pgadapter to driver
