+++ openmeter/credit/postgresdriver/repo.go
@@ -1,15 +1,15 @@
 package postgresadapter
 
 import (
-	"github.com/openmeterio/openmeter/internal/credit/postgresadapter"
+	postgresadapter "github.com/openmeterio/openmeter/internal/credit/postgresdriver"
 	"github.com/openmeterio/openmeter/openmeter/credit"
 	"github.com/openmeterio/openmeter/openmeter/entdb"
 )
 
-func NewPostgresBalanceSnapshotDBAdapter(db *entdb.DBClient) credit.BalanceSnapshotRepo {
+func NewPostgresBalanceSnapshotRepo(db *entdb.DBClient) credit.BalanceSnapshotRepo {
 	return postgresadapter.NewPostgresBalanceSnapshotRepo(db)
 }
 
-func NewPostgresGrantDBAdapter(db *entdb.DBClient) credit.GrantRepo {
+func NewPostgresGrantRepo(db *entdb.DBClient) credit.GrantRepo {
 	return postgresadapter.NewPostgresGrantRepo(db)
 }
